Phase 3: DB 스키마-엔티티 정합성 수정 및 빌드 환경 구성
- Stock.java 삭제, StockInfo.java로 통합 (동일 테이블 중복 엔티티 제거)
- StockInfo에 current_price, sector, volume 등 컬럼 추가, @Builder 적용
- StockRepository, PortfolioService: Stock -> StockInfo 기반으로 전환
- AssetUpdateScheduler SQL: trade_type -> buy_sell, 자기참조 JOIN 수정
- 7개 엔티티 @GeneratedValue(IDENTITY) 누락 추가 (Quiz, KnowTerm 등)
- Account nullable=false 완화, KnowTerm.userId Integer->Long 타입 수정
- Dockerfile COPY 경로 수정 (루트 기준으로 통일)
- .gitignore에 firebaseServiceAccountKey.json 추가

+++ BE/src/main/java/com/example/LAGO/service/DailyQuizService.java
@@ -83,7 +83,7 @@ public class DailyQuizService {
 
         DailySolved dailySolved = DailySolved.builder()
                 .solvedId(generateSolvedId())
-                .userId(userId)
+                .userId(userId.longValue())
                 .quizId(quizId)
                 .score(score)
                 .solvedAt(today)
@@ -171,7 +171,7 @@ public class DailyQuizService {
         } else {
             KnowTerm newKnowTerm = KnowTerm.builder()
                     .knowId(generateKnowId())
-                    .userId(userId)
+                    .userId(userId.longValue())
                     .termId(termId)
                     .correct(correct)
                     .build();
